+++ src/@locale/components/buttons/fullscreen.js
@@ -1,0 +1,15 @@
+const fullscreen =
+{
+    fr :
+    {
+        title   : 'Plein écran' ,
+        tooltip : 'Plein écran' ,
+    } ,
+    en :
+    {
+        title   : 'Fullscreen' ,
+        tooltip : 'Fullscreen' ,
+    }
+}
+
+export default fullscreen ;