+++ src/@locale/components/buttons/index.js
@@ -2,12 +2,14 @@ import add        from './add'
 import clear      from './clear'
 import copy       from './copy'
 import download   from './download'
+import fullscreen from './fullscreen'
 import less       from './less'
 import more       from './more'
 import refresh    from './refresh'
 import remove     from './remove'
 import revoke     from './revoke'
 import save       from './save'
+import theme      from './theme'
 
 const buttons =
 {
@@ -15,12 +17,14 @@ const buttons =
     clear ,
     copy ,
     download ,
+    fullscreen ,
     less ,
     more ,
     refresh ,
     remove ,
     revoke ,
     save ,
+    theme ,
 };
 
 export default buttons ;